+++ foxes/engines/__init__.py
@@ -1,3 +1,4 @@
 from .dask import DaskBaseEngine, XArrayEngine, DaskEngine, LocalClusterEngine, SlurmClusterEngine
 from .multiprocess import MultiprocessEngine
 from .numpy import NumpyEngine
+from .single_chunk import SingleChunkEngine